move unbatch serialization tests to kernel tests
